+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- All Stairs+ nodes can be now converted into microblocks
+
 ## [1.1.0] - 2017-10-04
 
 ### Added
@@ -26,7 +30,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Tin Block
   - Wool (all colors)
 - Other mods can now get a list of all the defined Stairs+ shapes
-- All Stairs+ nodes can be now converted into microblocks
 
 ## [1.0.0] - 2017-02-19